requirements.txt
djnago
pillow
djangorestframework
djangorestframework-simplejwt
django-cors-headers